+++ Code/Tests/mathtest/jan_math.cpp
@@ -588,12 +588,12 @@ int jan_main(int argc, char* argv[])
 		float r0 = 0.0f;
 		float r1 = 0.0f;
 		float r2 = 0.0f;
-		unsigned int acos_cycles = 0;
-		unsigned int bez_cycles = 0;
-		unsigned int table_cycles = 0;
-		unsigned int acos_sum = 0;
-		unsigned int bez_sum = 0;
-		unsigned int table_sum = 0;
+		uint64_t acos_cycles = 0;
+		uint64_t bez_cycles = 0;
+		uint64_t table_cycles = 0;
+		uint64_t acos_sum = 0;
+		uint64_t bez_sum = 0;
+		uint64_t table_sum = 0;
 		const int SAMPLE_COUNT = 20;
 
 		{
